#2 Consolidate new changes
